+++ es/identity.js
@@ -1,0 +1,9 @@
+/**
+ * Returns the first argument.
+ *
+ * @param  {*} value Value to return
+ * @return {*}       Value returned
+ */
+export default function identity( value ) {
+	return value;
+}