+++ rollup.cjs.config.js
@@ -1,11 +1,8 @@
 export default {
-	entry: 'es/rememo.js',
+	entry: 'es/index.js',
 	dest: 'rememo.js',
 	format: 'cjs',
 	external: [
-		'shallow-equal/arrays',
-		'moize',
-		'moize/lib/utils',
-		'moize/lib/Cache'
+		'moize'
 	]
 };