+++ rollup.iife.config.js
@@ -2,7 +2,7 @@ import resolve from 'rollup-plugin-node-resolve';
 import commonjs from 'rollup-plugin-commonjs';
 
 export default {
-	entry: 'es/rememo.js',
+	entry: 'es/index.js',
 	dest: 'dist/rememo.js',
 	moduleName: 'rememo',
 	format: 'iife',